didnt done the main

+++ dotNet5782_6400_5230/DAL/Client.cs
@@ -10,8 +10,8 @@ namespace IDAL
             public int ID { get; set; }
             public string name { get; set; }
             public int phoneNumber { get; set; }
-            public int longitude { get; set; }
-            public int latitude { get; set; }
+            public double longitude { get; set; }
+            public double latitude { get; set; }
             public override string ToString()
             {
                 return ("ID: " + ID + '\n' +
